taxshield-app/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.4
python-dotenv==1.0.1
openai==1.58.1
email-validator==2.2.0
